Add user card to unauthorized user
