+++ src/Components/Pages/Settings.js
@@ -22,8 +22,8 @@ import OpenInNewIcon from "@mui/icons-material/OpenInNew";
 import { defaultColorKey, colorOptions } from "../../constants";
 import { setAllData, deleteRequest, uploadColors } from "../../utils/requests";
 import { downloadData } from "../../utils/dataTransfer";
-import { signOutGoogle } from "../../utils/googleAuth";
 import ImportFileButton from "../Utils/ImportFileButton";
+import UserCard from "../Utils/UserCard";
 import {
   AddAlertContext,
   DataPathsContext,
@@ -61,10 +61,6 @@ function Settings(props) {
     downloadData(data);
   };
 
-  const handleLogout = () => {
-    signOutGoogle(addAlert);
-  };
-
   const renderAppCard = () => {
     return (
       <Box sx={{ width: "95%" }}>
@@ -94,31 +90,6 @@ function Settings(props) {
               .
             </Typography>
           </CardContent>
-        </Card>
-      </Box>
-    );
-  };
-
-  const renderUserCard = () => {
-    if (!user) {
-      return;
-    }
-    return (
-      <Box sx={{ width: "95%" }}>
-        <Card variant="outlined">
-          <CardContent>
-            <Typography variant="h6" color="text.secondary" gutterBottom>
-              Current user
-            </Typography>
-            <Typography>
-              You are currently logged in as: <strong>{user.email}</strong>
-            </Typography>
-          </CardContent>
-          <CardActions sx={{ justifyContent: "flex-end" }}>
-            <Button color="secondary" variant="outlined" onClick={handleLogout}>
-              Logout
-            </Button>
-          </CardActions>
         </Card>
       </Box>
     );
@@ -284,7 +255,7 @@ function Settings(props) {
         Settings
       </Typography>
       <Stack sx={{ paddingTop: "15px" }} spacing={3} alignItems="center">
-        {renderUserCard()}
+        <UserCard user={user} addAlert={addAlert} />
         {renderColorCard()}
         {renderAppCard()}
         {renderDownloadData()}
